refactored Application controllers
