+++ app/Http/Controllers/Auth/Application/ApplicationAdminController.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Http\Controllers\Auth\Application;
+
+use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
+use App\Models\Users\ApplicationForm;
+
+class ApplicationAdminController extends Controller
+{
+}
